[android] #2588 - Moving InfoWindow classes to annotations package
