+++ android/java/MapboxGLAndroidSDK/src/main/res/layout/infowindow.xml
@@ -44,7 +44,7 @@
             android:visibility="gone"/>
     </LinearLayout>
 
-    <com.mapbox.mapboxsdk.views.InfoWindowTipView
+    <com.mapbox.mapboxsdk.annotations.InfoWindowTipView
         android:layout_width="fill_parent"
         android:layout_height="10dp"/>
